Edit modal save button should be disabled when info app is visible
Ideally the save button shouldn't be disabled, but currently the handleConfirm method relies on finding the different inputs field in the DOM, but they are only part of the DOM when the settings app is activated, not the info app. Ideally this should be fixed, but for now, making the save button disabled prevents the user from saving when the info app is activated.